# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AVIV Clan+ API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clans/{clanId}/members:
    get:
      operationId: listClanMembers
      tags: [members]
      summary: List clan members
      parameters:
        - $ref: "#/components/parameters/ClanId"
      responses:
        "200":
          description: Members list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClanMember"
  /clans/{clanId}/members/me:
    patch:
      operationId: updateMyMembership
      tags: [members]
      summary: Update my membership settings (e.g. silence this clan)
      parameters:
        - $ref: "#/components/parameters/ClanId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MembershipUpdate"
      responses:
        "200":
          description: Updated membership
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClanMember"
        "404":
          description: Not a member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: leaveClан
      tags: [members]
      summary: Leave a clan
      parameters:
        - $ref: "#/components/parameters/ClanId"
      responses:
        "204":
          description: Left
  /clans/{clanId}/members/{userId}:
    delete:
      operationId: kickMember
      tags: [members]
      summary: Remove a member from the clan (leader only)
      parameters:
        - $ref: "#/components/parameters/ClanId"
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
        "403":
          description: Not the clan leader
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Join requests ─────────────────────────────────────────────────────────
  /clans/{clanId}/join-requests:
    post:
      operationId: requestToJoin
      tags: [joinRequests]
      summary: Request to join a clan
      parameters:
        - $ref: "#/components/parameters/ClanId"
      responses:
        "201":
          description: Request submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JoinRequest"
        "409":
          description: Already a member or request already pending
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: listJoinRequests
      tags: [joinRequests]
      summary: List pending join requests (leader only)
      parameters:
        - $ref: "#/components/parameters/ClanId"
      responses:
        "200":
          description: Pending requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JoinRequest"
  /clans/{clanId}/join-requests/{requestId}:
    patch:
      operationId: resolveJoinRequest
      tags: [joinRequests]
      summary: Approve or reject a join request (leader only)
      parameters:
        - $ref: "#/components/parameters/ClanId"
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinRequestResolution"
      responses:
        "200":
          description: Resolved request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JoinRequest"

  # ── Invite ────────────────────────────────────────────────────────────────
  /clans/{clanId}/invite-token:
    get:
      operationId: getInviteToken
      tags: [invite]
      summary: Get or create an invite token for the clan (leader only)
      parameters:
        - $ref: "#/components/parameters/ClanId"
      responses:
        "200":
          description: Invite token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteTokenInfo"
    post:
      operationId: regenerateInviteToken
      tags: [invite]
      summary: Regenerate the invite token (leader only)
      parameters:
        - $ref: "#/components/parameters/ClanId"
      responses:
        "200":
          description: New invite token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteTokenInfo"
  /invite/{token}:
    get:
      operationId: getInviteInfo
      tags: [invite]
      summary: Get clan info from an invite token (no auth required)
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Clan preview for the invite
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteClanPreview"
        "404":
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /invite/{token}/join:
    post:
      operationId: joinViaInvite
      tags: [invite]
      summary: Join a clan via invite link (auto-approved)
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Joined
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClanMember"
        "409":
          description: Already a member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Discord ───────────────────────────────────────────────────────────────
...
  /clans/{clanId}/push-subscribe:
    post:
      operationId: pushSubscribe
      tags: [push]
      summary: Subscribe to web push notifications for a clan
      parameters:
        - $ref: "#/components/parameters/ClanId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PushSubscriptionInput"
      responses:
        "201":
          description: Subscribed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
    delete:
      operationId: pushUnsubscribe
      tags: [push]
      summary: Remove web push subscription for a clan
      parameters:
        - $ref: "#/components/parameters/ClanId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PushUnsubscribeInput"
      responses:
        "204":
          description: Unsubscribed
  /clans/{clanId}/push-vapid:
    get:
      operationId: getVapidKey
      tags: [push]
      summary: Get the VAPID public key for this clan push channel
      parameters:
        - $ref: "#/components/parameters/ClanId"
      responses:
        "200":
          description: VAPID public key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VapidKeyInfo"
components:
  parameters:
    ClanId:
      name: clanId
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]
...
    ClanMember:
      type: object
      properties:
        userId:
          type: integer
        clanId:
          type: integer
        username:
          type: string
        avatar:
          type: ["string", "null"]
        role:
          type: string
          enum: [leader, member]
        silenced:
          type: boolean
        joinedAt:
          type: string
      required: [userId, clanId, username, avatar, role, silenced, joinedAt]
    MembershipUpdate:
      type: object
      properties:
        silenced:
          type: boolean
...
    JoinRequest:
      type: object
      properties:
        id:
          type: integer
        clanId:
          type: integer
        userId:
          type: integer
        username:
          type: string
        avatar:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        createdAt:
          type: string
      required: [id, clanId, userId, username, avatar, status, createdAt]
    JoinRequestResolution:
      type: object
      properties:
        action:
          type: string
          enum: [approve, reject]
      required: [action]
    InviteTokenInfo:
      type: object
      properties:
        token:
          type: string
        url:
          type: string
      required: [token, url]
    InviteClanPreview:
      type: object
      properties:
        clanId:
          type: integer
        name:
          type: string
        imageUrl:
          type: ["string", "null"]
        memberCount:
          type: integer
        leaderUsername:
          type: string
      required: [clanId, name, imageUrl, memberCount, leaderUsername]
...
    PushSubscriptionInput:
      type: object
      properties:
        endpoint:
          type: string
        p256dh:
          type: string
        auth:
          type: string
      required: [endpoint, p256dh, auth]
    PushUnsubscribeInput:
      type: object
      properties:
        endpoint:
          type: string
      required: [endpoint]
    VapidKeyInfo:
      type: object
      properties:
        publicKey:
          type: string
      required: [publicKey]