# lib/api-zod/src/generated/types/joinRequestResolution.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AVIV Clan+ API
 * OpenAPI spec version: 0.1.0
 */
import type { JoinRequestResolutionAction } from './joinRequestResolutionAction';

export interface JoinRequestResolution {
  action: JoinRequestResolutionAction;
}